Listo

+++ Clase 2./Expresiones.java
@@ -1,8 +1,8 @@
 public class Expresiones {
 public static void main(String[]args){
- int A = 4;
- int B = 5;
- int C = 1;
+ double A = 4;
+ double B = 5;
+ double C = 1;
       System.out.println(B * A - Math.pow(B,2) / 4 * C);
 }
 }
